_wapos/requirements.txt
alembic==1.16.1
astroid==3.3.10
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dill==0.4.0
dotenv==0.9.9
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
h11==0.16.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pylint==3.3.7
pytest==8.4.0
python-dotenv==1.1.0
python-engineio==4.12.1
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.41
tomlkit==0.13.2
typing_extensions==4.14.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1